refactor defaults, patterns, and types into an ssvc.utils sub-package
