+++ src/ssvc/namespaces.py
@@ -23,109 +23,10 @@ for SSVC and provides a method to validate namespace values.
 #  subject to its own license.
 #  DM24-0278
 
-import re
 from enum import StrEnum, auto
-from typing import Annotated
 
-from pydantic import Field
-
-MIN_NS_LENGTH = 3
-MAX_NS_LENGTH = 1000
-NS_LENGTH_INTERVAL = MAX_NS_LENGTH - MIN_NS_LENGTH
-
-# from https://docs.oasis-open.org/csaf/csaf/v2.0/os/csaf-v2.0-os.html
-BCP_47_PATTERN = r"(([A-Za-z]{2,3}(-[A-Za-z]{3}(-[A-Za-z]{3}){0,2})?|[A-Za-z]{4,8})(-[A-Za-z]{4})?(-([A-Za-z]{2}|[0-9]{3}))?(-([A-Za-z0-9]{5,8}|[0-9][A-Za-z0-9]{3}))*(-[A-WY-Za-wy-z0-9](-[A-Za-z0-9]{2,8})+)*(-[Xx](-[A-Za-z0-9]{1,8})+)?|[Xx](-[A-Za-z0-9]{1,8})+|[Ii]-[Dd][Ee][Ff][Aa][Uu][Ll][Tt]|[Ii]-[Mm][Ii][Nn][Gg][Oo])"
-"""A regular expression pattern for BCP-47 language tags."""
-
-LENGTH_CHECK_PATTERN = rf"(?=.{{{MIN_NS_LENGTH},{MAX_NS_LENGTH}}}$)"
-"""Ensures the string is between MIN_NS_LENGTH and MAX_NS_LENGTH characters long."""
-
-# Base namespace part (before any extensions) allows . and - with restrictions
-BASE_PATTERN = (
-    r"(?!.*[.-]{2,})"  # no consecutive separators
-    r"[a-z][a-z0-9]+"  # first part starts with a letter, followed by one or more alphanumeric characters
-    r"(?:[.-][a-z0-9]+)*"  # remaining parts can have alphanumeric characters and single . or - separators
-)
-"""The base pattern for namespaces, which must start with a letter and contain at least 3 alphanumeric characters."""
-
-X_PFX = "x_"
-"""The prefix for extension namespaces. Extension namespaces must start with this prefix."""
-
-EXPERIMENTAL_BASE = rf"{X_PFX}{BASE_PATTERN}"
-f"""The base pattern for experimental namespaces, which must start with the {X_PFX} prefix, 
-followed by a string matching the base pattern."""
-
-BASE_NS_PATTERN = rf"({EXPERIMENTAL_BASE}|{BASE_PATTERN})"
-"""The complete base namespace pattern, which allows for experimental namespaces."""
-
-# Extension segment pattern (alphanumeric + limited punctuation, no consecutive punctuation, ends with alphanumeric)
-EXT_SEGMENT_PATTERN = (
-    r"(?!.*[.-]{2,})"  # no consecutive separators
-    r"[a-zA-Z][a-zA-Z0-9]*"  # first part starts with a letter, followed by one or more alphanumeric characters
-    r"(?:[.-][a-zA-Z0-9]+)*"  # remaining parts can have alphanumeric characters and single ., -, / separators
-)
-"""The pattern for extension segments in namespaces, which must start with a letter and contain alphanumeric characters or
-limited punctuation characters (., -), with no consecutive punctuation characters allowed."""
-
-# Language extension pattern (BCP-47 or empty for //)
-LANG_EXT_PATTERN = rf"(/({BCP_47_PATTERN})/|//)"
-"""The pattern for the first extension segment, which must be either a valid BCP-47 tag or empty (//)."""
-
-# Subsequent extension segments
-SUBSEQUENT_EXT_PATTERN = rf"{EXT_SEGMENT_PATTERN}(?:/{EXT_SEGMENT_PATTERN})*"
-"""The pattern for subsequent extension segments, which must follow the rules for extension segments, delimited by slashes (/)."""
-
-# Complete pattern with length validation
-NS_PATTERN = re.compile(
-    rf"^{LENGTH_CHECK_PATTERN}({BASE_NS_PATTERN})({LANG_EXT_PATTERN}{SUBSEQUENT_EXT_PATTERN})?$"
-)
-f"""The full regular expression pattern for validating namespaces.
-
-!!! note "Length Requirements"
-
-    - Namespaces must be between {MIN_NS_LENGTH} and {MAX_NS_LENGTH} characters long.
-
-!!! note "Base Namespace Requirements"
-    
-    - Must start with a lowercase letter
-    - Must contain at least 3 total characters in the base part (after the optional experimental/private prefix)
-    - Must contain only lowercase letters, numbers, dots (`.`), and hyphens (`-`)
-    - Must not contain consecutive dots or hyphens (no `..`, `--`, `.-`, `-.`, `---`, etc.)
-    - May optionally start with the experimental/private prefix `{X_PFX}`.
-    
-!!! note "Extension Requirements (Optional)"
-    
-    - Extensions are optional
-    - Extensions must be delineated by slashes (`/`)
-    - If any extension segments are present, the following rules apply:
-    - The first extension segment, must be a valid BCP-47 language tag or empty (i.e., `//`).
-    - Subsequent extension segments:
-        - must start with a letter (upper or lowercase)
-        - may contain letters, numbers, dots (`.`), and hyphens (`-`)
-        - must not start or end with a dot or hyphen
-        - must not contain consecutive dots or hyphens (no `..`, `--`, `.-`, `-.`, `---`, etc.)
-        - are separated by single forward slashes (`/`)
-    - multiple extension segments are allowed
-    
-"""
-
-NamespaceString = Annotated[
-    str,
-    Field(
-        description="The namespace of the SSVC object.",
-        examples=[
-            "ssvc",
-            "cisa",
-            "x_com.example//private",
-            "com.example//some-extension",
-            "ssvc/de-DE/example.organization/reference-arch-1",
-        ],
-        pattern=NS_PATTERN,
-        min_length=MIN_NS_LENGTH,
-        max_length=MAX_NS_LENGTH,
-    ),
-]
-"""A string datatype for namespace values, for use in Pydantic models."""
+from ssvc.utils.defaults import MAX_NS_LENGTH, MIN_NS_LENGTH, X_PFX
+from ssvc.utils.patterns import NS_PATTERN
 
 
 class NameSpace(StrEnum):
